+++ packages/next-core/src/index.ts
@@ -1,6 +1,10 @@
 export { handleApiRequest } from "./requestHandler.js";
 export * from "@fistware/http-core";
-export { logger } from "./lib/logger.js";
+export {
+  configureNextCoreLogger,
+  logger,
+  type ConfigureNextCoreLoggerOptions,
+} from "./lib/logger.js";
 export { HandleApiRequestOptions } from "./lib/interfaces/HandleApiRequestOptions.js";
 export { NextRequestParts } from "./lib/interfaces/NextRequestParts.js";
 export { extractRequestParts } from "./lib/utils/utils.js";